WIP: First draft of adding IR printing functionality of the PassManager
So far, a `Print` pass has been added every time the OpPassManager::nest() is called.
The `Print` pass presented a good starting point for this issue, since it had
implemented the basic building blocks for IR printing.

What is currently missing is:
- Configuring the IRPrintingConfig struct. At the moment, the IRPrintingConfig
This, in turn, means that regardless of if there have been modifications to the
IR or not the IR is *always* printed. The issue states that this should be
configurable. I assume that these configurations should come from the
Session/Context struct (although I don't think there's a 1 to 1 correspondance).
    - Realted to this, making printing granular is also missing. At the moment,
    the Print Pass is applied irregardless of the type. The issue also states
    that it should be able to *only print IR when a Pass modified the IR*.
- One interesting note is that the PassManager has some functionality related to
IR printing, although after some initial attempts I found it simpler to add that
logic in the OpPassManager itself.

Signed-off-by: Tomas Fabrizio Orsi <tomas.orsi@lambdaclass.com>